claude(scanning-git-for-tils): move tests into subfolders
